Refactored my main class. Split it into 5 smaller classes which will be (hopefully) easier to maintain.
